backend/requirements.txt
Django==5.2.10
djangorestframework
django-cors-headers
requests
slack_sdk
